+++ src/components/authentication/authRoute.js
@@ -4,7 +4,7 @@ import { Route, Redirect } from 'react-router-dom'
 class AuthRoute extends Component {
 
   render() {
-    const logged = localStorage.getItem('loggedUser') ? true : false
+    const logged = localStorage.getItem('loggedUser');
     const Comp = this.props.component;
     return (
       <Route render={props => (
