+++ src/components/authentication/privateRoute.js
@@ -1,14 +1,14 @@
 import React, { Component } from 'react';
 import { Route, Redirect } from 'react-router-dom'
-import ApiService from '../customComponents/apiService';
+
 class PrivateRoute extends Component {
 
   render() {
-    const logged = localStorage.getItem('loggedUser') ? true : false
+    const logged = localStorage.getItem('loggedUser')
     const Comp = this.props.component;
     return (
       <Route render={props => (
-        logged ? ApiService.resetToken() && <Comp {...this.props} />
+        logged ? <Comp {...this.props} />
          : (
         <Redirect to={{
           pathname: '/login',
